move tests into indexer root folder
